refactor: move FlightType to types/ folder
